refactor: move render_args_as_flags to shell module and update its usage across the codebase
